refactor: final verification and shared assets organization
