+++ src/components/StepperControl/StepperControl.tsx
@@ -1,6 +1,6 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import styles from "./StepperControl.module.css";
-import shared from "../shared.module.css";
+import shared from "../shared/shared.module.css";
 
 const stepperControlVariants = cva(styles["stepper-control"], {
   variants: {
