+++ src/components/shared/shared.module.css
